requirements.txt
scrapy==2.5.0
requests==2.26.0
beautifulsoup4==4.10.0
Twisted>=22.10.0
pytz==2022.1
schedule==1.2.1
psycopg2-binary==2.9.5
